inode_context: fix uint32_t overflow

+++ src/context/inode_context.c
@@ -324,5 +324,5 @@ squash_inode_ref_to_block(
 }
 uint64_t
 squash_inode_ref_from_block(uint32_t block_index, uint16_t offset) {
-	return (block_index << 16) | offset;
+	return ((uint64_t)block_index << 16) | offset;
 }
